+++ Blog.Core.Model/Models/AccessTrendLog.cs
@@ -6,7 +6,7 @@ namespace Blog.Core.Model.Models
     /// <summary>
     /// 用户访问趋势日志
     /// </summary>
-    public class AccessTrendLog : RootEntityTkey<int>
+    public class AccessTrendLog : RootEntityTkey<long>
     {
         /// <summary>
         /// 用户
